Remove inessesry migrations

+++ database/migrations/2025_05_31_113911_create_article_palette_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(ArticleStock::class);
             $table->foreignId('palette_id')->constrained()->onDelete('cascade');
-            $table->integer('quantity')->default(0);
+            $table->float('quantity')->default(0.0);
             $table->timestamps();
         });
     }
